requirements.txt
openai
python-dotenv
streamlit
scikit-learn
numpy
pandas
spacy
networkx
matplotlib
sentence-transformers
torch
pymupdf
pyvis
